refact : service, controller에서 @RequiredArgsConstructor로 생성자 대체

+++ src/main/java/com/billit/investment/repository/InvestStatusRepository.java
@@ -3,6 +3,6 @@ package com.billit.investment.repository;
 import com.billit.investment.domain.InvestStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InvestStatusRepository extends JpaRepository<InvestStatus, Long> {
+public interface InvestStatusRepository extends JpaRepository<InvestStatus, Integer> {
 }
